Separate YAML handling from Session
