+++ tests/test_sessionfile.py
@@ -14,7 +14,8 @@ from rflx.expression import (
     VariableDeclaration,
 )
 from rflx.identifier import ID
-from rflx.session import Session, SessionFile, State, Transition
+from rflx.session import Channel, Session, State, Transition
+from rflx.sessionfile import SessionFile
 from rflx.statement import Assignment
 
 
@@ -632,3 +633,130 @@ def test_session_with_renames() -> None:
         },
     )
     assert f.sessions[0] == expected
+
+
+def test_channels() -> None:
+    f = SessionFile()
+    f.parse_string(
+        "session",
+        """
+            channels:
+                - name: Channel1_Read_Write
+                  mode: Read_Write
+                - name: Channel2_Read
+                  mode: Read
+                - name: Channel3_Write
+                  mode: Write
+            initial: START
+            final: END
+            states:
+              - name: START
+                transitions:
+                  - target: END
+                variables:
+                  - "Local : Boolean"
+                actions:
+                  - Local := Write(Channel1_Read_Write, Read(Channel2_Read))
+                  - Local := Write(Channel3_Write, Local)
+              - name: END
+        """,
+    )
+    expected = Session(
+        name="session",
+        initial="START",
+        final="END",
+        states=[
+            State(
+                name="START",
+                transitions=[Transition(target="END")],
+                declarations={ID("Local"): VariableDeclaration("Boolean")},
+                actions=[
+                    Assignment(
+                        "Local",
+                        SubprogramCall(
+                            "Write",
+                            [
+                                Variable("Channel1_Read_Write"),
+                                SubprogramCall("Read", [Variable("Channel2_Read")]),
+                            ],
+                        ),
+                    ),
+                    Assignment(
+                        "Local",
+                        SubprogramCall("Write", [Variable("Channel3_Write"), Variable("Local")],),
+                    ),
+                ],
+            ),
+            State(name="END"),
+        ],
+        declarations={
+            "Channel1_Read_Write": Channel(read=True, write=True),
+            "Channel2_Read": Channel(read=True, write=False),
+            "Channel3_Write": Channel(read=False, write=True),
+        },
+    )
+    assert f.sessions[0] == expected
+
+
+def test_channel_with_invalid_mode() -> None:
+    with pytest.raises(
+        RecordFluxError,
+        match="^session: error: channel Channel1_Read_Write has invalid mode Invalid",
+    ):
+        SessionFile().parse_string(
+            "session",
+            """
+                channels:
+                    - name: Channel1_Read_Write
+                      mode: Invalid
+                initial: START
+                final: END
+                states:
+                  - name: START
+                    transitions:
+                      - target: END
+                    variables:
+                      - "Local : Boolean"
+                    actions:
+                      - Local := Read(Channel1_Read_Write)
+                  - name: END
+            """,
+        )
+
+
+def test_channel_without_name() -> None:
+    with pytest.raises(RecordFluxError, match="^session: error: channel 0 has no name"):
+        SessionFile().parse_string(
+            "session",
+            """
+                channels:
+                    - mode: Read_Write
+                initial: START
+                final: END
+                states:
+                  - name: START
+                    transitions:
+                      - target: END
+                  - name: END
+            """,
+        )
+
+
+def test_channel_without_mode() -> None:
+    with pytest.raises(
+        RecordFluxError, match="^session: error: channel Channel_Without_Mode has no mode"
+    ):
+        SessionFile().parse_string(
+            "session",
+            """
+                channels:
+                    - name: Channel_Without_Mode
+                initial: START
+                final: END
+                states:
+                  - name: START
+                    transitions:
+                      - target: END
+                  - name: END
+            """,
+        )
